Solution2
