+++ src/index.tsx
@@ -1,9 +1,13 @@
-import { createRoot } from 'react-dom/client';
-import { HashRouter } from 'react-router-dom';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import { App } from './App';
+import './App.scss';
 
-createRoot(document.getElementById('root') as HTMLElement).render(
-  <HashRouter>
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+root.render(
+  <React.StrictMode>
     <App />
-  </HashRouter>,
+  </React.StrictMode>
 );
