+++ src/pages/HomePage.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const HomePage: React.FC = () => (
+  <div className="section">
+    <div className="container">
+      <h1 className="title">Home page</h1>
+    </div>
+  </div>
+);
