+++ src/pages/PageNotFound.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const PageNotFound: React.FC = () => (
+  <div className="section">
+    <div className="container">
+      <h1 className="title">Page not found</h1>
+    </div>
+  </div>
+);
