+++ src/pages/TabsPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Tabs } from '../components/Tabs';
+
+export const TabsPage: React.FC = () => (
+  <div className="section">
+    <div className="container">
+      <h1 className="title">Tabs page</h1>
+      <Tabs />
+    </div>
+  </div>
+);
